+++ netpyne/cell.py
@@ -615,7 +615,7 @@ class Cell (object):
                 
         else:
             if sim.cfg.verbose: print('Adding exotic stim (NeuroML2 based?): %s'% params['type'])   
-            stim = getattr(h, params['type'])(sec['hSection'](params['loc']))
+            stim = getattr(h, params['type'])(sec['hSec'](params['loc']))
             stimParams = {k:v for k,v in params.iteritems() if k not in ['type', 'source', 'loc', 'sec', 'label']}
             stringParams = ''
             for stimParamName, stimParamValue in stimParams.iteritems(): # set mechanism internal params
